+++ catalog/app/components/Preview/types.js
@@ -37,6 +37,7 @@ export const PreviewData = tagged([
   'Text', // { head: string, tail: string, lang: string, highlighted: { head: string, tail: string }, ...PreviewStatus }
   'Vcf', // { meta: string[], header: string[], body: string[][], variants: string[], ...PreviewStatus }
   'Vega', // { spec: Object }
+  'Video', // { src: string }
   'Voila', // { src: string }
 ])