+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@
 * [Added] Set height for `quilt_summarize.json` files ([#2474](https://github.com/quiltdata/quilt/pull/2474))
 * [Added] Add a "transcode" lambda for previewing video files ([#2366](https://github.com/quiltdata/quilt/pull/2366/))
 * [Added] Add object-level metadata editor and move package metadata editor to popup ([#2510](https://github.com/quiltdata/quilt/pull/2510/))
+* [Added] Video previews ([#2540](https://github.com/quiltdata/quilt/pull/2540))
 
 # 3.6.0 - 2021-10-15
 ## Python API